http://127.0.0.1:3000/index.html#home
Build a personal portfolio website that showcases your skills, projects and accomplishments as a web developer. Design an attractive and visually appealing layout that captures visitors attention. Include a home page with a captivating headline, professional photo and summary of your skills. provide an "about me" section with a detailed background , education and experience.